move constrain fn to auth_base and share with flat
