+++ config/auth_vpopmaild.ini
@@ -1,9 +1,9 @@
 [main]
 host=127.0.0.6
 port=89
-;sysadmin=postmaster@example.com:sekret
-constrain_sender=true
+; sysadmin=postmaster@example.com:sekret
+; constrain_sender=true
 
 [example.com]
 host=127.0.0.10
-;sysadmin=postmaster@example.com:sekret
+; sysadmin=postmaster@example.com:sekret
